[feat] 체력바 플레이어와 동기화 완료

+++ Assets/Scripts/Monster/MonsterStatusSO.cs
@@ -7,17 +7,17 @@ public abstract class MonsterStatusSO : ScriptableObject
     [Header("몬스터 기본 정보")]
     public int ID;
     public string Name;
-    public int MaxHP;
+    public uint MaxHP;
     public GameObject Prefab;
     //To-Do 아이템 클래스 추가
     //public List<> DropItem;
 
     [Header("몬스터 전투 관련")]
-    public float BiteDamage;
-    public float TaileDamage;
-    public float ChargeDamage;
-    public float FlyDamage;
-    public float Defense;
+    public uint BiteDamage;
+    public uint TaileDamage;
+    public uint ChargeDamage;
+    public uint FlyDamage;
+    public uint Defense;
     public float AttackCoolTime;
     public float DetectRange;
     public float AttackRange;
